Forecasting-Service-main/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
joblib
pandas
scikit-learn
prophet